+++ src/id_translation/__init__.py
@@ -30,7 +30,7 @@ __all__ = [
     "__version__",  # Make MyPy happy
 ]
 
-__version__ = "0.15.0.dev1"
+__version__ = "0.15.1"
 
 _logging.getLogger(__name__).addHandler(_logging.NullHandler())
 del _logging