+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.15.1] - 2025-07-20
+
 ### Added
 - New convenience method `Translator.initialize_sources()`.
 
@@ -436,7 +438,8 @@ cookiecutter template.
 - Switch to relative imports.
 - Fix some intersphinx issues.
 
-[Unreleased]: https://github.com/rsundqvist/id-translation/compare/v0.15.0...HEAD
+[Unreleased]: https://github.com/rsundqvist/id-translation/compare/v0.15.1...HEAD
+[0.15.1]: https://github.com/rsundqvist/id-translation/compare/v0.15.0...v0.15.1
 [0.15.0]: https://github.com/rsundqvist/id-translation/compare/v0.14.1...v0.15.0
 [0.14.1]: https://github.com/rsundqvist/id-translation/compare/v0.14.0...v0.14.1
 [0.14.0]: https://github.com/rsundqvist/id-translation/compare/v0.13.0...v0.14.0